server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
email-validator==2.1.0
python-jose[cryptography]
passlib[bcrypt]
python-dotenv

